# utils/swagger.json
{
    "info": {
        "description": "API backend para aplica\u00e7\u00e3o",
        "version": "1.0.0-oas3",
        "contact": {
            "email": "raynan.serafim@gmail.com"
        },
        "title": "Backend"
    },
    "servers": [
        {
            "description": "Server local",
            "url": "http://localhost:5000"
        }
    ],
...
    "paths": {
...
        "/acesso/view/{query_id}": {
            "get": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Busca o registro do banco se ele existir",
                "parameters": [
                    {
                        "in": "path",
                        "name": "query_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "description": "Identifica\u00e7\u00e3o \u00fanica do registro"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcessoModel"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Ocorreu um erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/acesso/all": {
            "get": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Busca lista de registro existentes no banco",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "integer"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AcessoModel"
                                            }
                                        }
                                    },
                                    "required": [
                                        "count",
                                        "items"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/endereco/view/{query_id}": {
            "get": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Busca o registro do banco se ele existir",
                "parameters": [
                    {
                        "in": "path",
                        "name": "query_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "description": "Identifica\u00e7\u00e3o \u00fanica do registro"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EnderecoModel"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Ocorreu um erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/municipio/list": {
            "get": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Busca lista de registro existentes no banco",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "integer"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EnderecoModel"
                                            }
                                        }
                                    },
                                    "required": [
                                        "count",
                                        "items"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/endereco/update/{query_id}": {
            "put": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Edita um registro",
                "parameters": [
                    {
                        "in": "path",
                        "name": "query_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "description": "Identifica\u00e7\u00e3o \u00fanica do registro"
                    }
                ],
                "requestBody": {
                    "description": "Dados necess\u00e1rios para a edi\u00e7\u00e3o do registro",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EnderecoModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu um erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "openapi": "3.0.2",
    "components": {
        "schemas": {
            "AcessoModel": {
                "title": "AcessoModel",
                "type": "object",
                "properties": {
                    "nome": {
                        "title": "Nome",
                        "maxLength": 255,
                        "type": "string"
                    }
                },
                "required": [
                    "nome"
                ]
            },
            "EnderecoModel": {
                "title": "EnderecoModel",
                "type": "object",
                "properties": {
                    "cep": {
                        "title": "Cep",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "rua": {
                        "title": "Rua",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "numero": {
                        "title": "Numero",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "bairro": {
                        "title": "Bairro",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "complemento": {
                        "title": "Complemento",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "municipio_id": {
                        "title": "Municipio Id",
                        "type": "integer"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}